Added Singleton MusicManager logic and added commands for music playback
- The pause command is not implemented yet.
The MusicManager being a Singleton, we bypass the need to put it in the buffer and mess with memory management : java will handle the players Map itself.
This also means that there is two "buffer" and trying to clear out all the data from a single guild will be a chore.