Fix killing commands
The kill() method now calls the stopAction() method and if the return value is true, it gives the command a "dead" state which then returns that state.

This change makes it so the stopAction() method is like a "code to do before killing this command", but also needs to be implemented to actually kill a command (making it "killable").