Generalize all Music classes to allow usage outside of commands
This allows for different events to also deal with the players and start new ones if this is the case we might want to do in the future.

There are still methods that uses the commands, but these are there only for legacy purposes and facilitate the usage in actual commands.