Tweak CommandSetting to change settings for every TextChannel
When using the Guild parameter (-g or --guild), previously, it would've changed the settings in the Guild level, but if settings were already defined in a TextChannel level, that setting would not be applied to the TextChannel level. Now, using the Guild level, everything in the Guild is updated to match that setting preference.

Individual Settings can be configured to only run for guilds or prevent being modified on the guild level.